Add specialized fixed-width integer / float DataType subclasses
Each common width gets its own concrete class — ``Int8Type``,
``Int16Type``, ``Int32Type``, ``Int64Type``, the matching ``UIntXType``
quartet, plus ``Float16Type`` / ``Float32Type`` / ``Float64Type`` —
backed by dedicated ``DataTypeId`` values (``INT8`` … ``UINT64``,
``FLOAT16`` … ``FLOAT64``).

``IntegerType.__new__`` and ``FloatingPointType.__new__`` redirect to
the registered specialized class for ``(byte_size, signed)`` /
``byte_size``, falling back to the abstract base for unusual widths
(``int128``, ``hugeint``, ``None``). The parser routes sized aliases
(``int32``, ``uint8``, ``bigint``, ``double``, ``f16``, ``bfloat16``,
…) to the specialized ids so ``parse_data_type`` carries signedness +
width without re-scanning metadata, and ``from_parsed`` reconstructs
the matching class from the id alone.

``to_dict`` now serializes ``signed`` so unsigned types round-trip
through generic-id payloads, and ``NumericType.merge_with`` treats any
two integer subtypes (or any two float subtypes) as same-kind so
``Int32.merge_with(Int64)`` collapses by width / signedness exactly as
the pre-split ``IntegerType`` did. Old-format dicts (``id=3 INTEGER``
with ``byte_size`` + ``signed``) keep resolving to the right
specialized class via the ``__new__`` redirect.

+++ python/tests/test_yggdrasil/test_data/test_schema.py
@@ -528,7 +528,7 @@ class TestAutotag:
         out = s.autotag(tags={"layer": "silver"})
 
         id_tags = out["user_id"].tags or {}
-        assert id_tags[b"type_name"] == b"integer"
+        assert id_tags[b"type_name"] == b"int64"
         assert id_tags[b"signed"] == b"true"
         assert id_tags[b"nullable"] == b"false"
         assert id_tags[b"primary_key"] == b"true"
